markdown features + styling eksempler og opstart på brug af utils, for bedre struktur
