+++ hello_app/templates/mdnote.html
@@ -38,13 +38,13 @@ View Note
 </div>
  
 <p class="view-timestamps">Oprettet: {{ note.created.strftime("%Y-%m-%d %H:%M") }} / 
-Redigeret: {{ note.lastEdited.strftime("%Y-%m-%d %H:%M") }}</p>
+Redigeret: {{ note.lastedited.strftime("%Y-%m-%d %H:%M") }}</p>
 <!-- Actions button-->
 <div align="center">
-    <a href="{{ url_for('edit', id=note.id) }}" class="icon-link">
+    <a href="{{ url_for('edit', id=note.noteID) }}" class="icon-link">
         <i class="fa-solid fa-wrench"></i>
     </a>
-    <a href="{{ url_for('delete_note', id=note.id) }}" class="icon-link">
+    <a href="{{ url_for('delete_note', id=note.noteID) }}" class="icon-link">
         <i class="fa-solid fa-trash-can"></i>
     </a>
